webapp/requirements.txt
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
selenium==3.8.0
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
